Refactoring tests
Creating a generic test case object to encapsulate database creation and
initialization. Added other utility functions to module too i.e.
get_record ti return a sample record.
